requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
altair=5.2.0=pypi_0
annotated-types=0.6.0=pypi_0
anyio=4.1.0=pypi_0
appnope=0.1.3=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=2.4.1=pypi_0
async-lru=2.0.4=pypi_0
attrs=23.1.0=pypi_0
aws-c-auth=0.7.8=hcd8a21d_2
aws-c-cal=0.6.9=had988b7_2
aws-c-common=0.9.10=h10d778d_0
aws-c-compression=0.2.17=hb45f1eb_7
aws-c-event-stream=0.3.2=hbc2660c_8
aws-c-http=0.7.14=h1fa4523_3
aws-c-io=0.13.36=h3728bb0_2
aws-c-mqtt=0.9.10=h2150435_2
aws-c-s3=0.4.4=h960c460_0
aws-c-sdkutils=0.1.13=hb45f1eb_0
aws-checksums=0.1.17=hb45f1eb_6
aws-crt-cpp=0.24.11=hc04ff71_2
aws-sdk-cpp=1.11.210=h7d17a91_4
babel=2.13.1=pypi_0
beautifulsoup4=4.12.2=pypi_0
bleach=6.1.0=pypi_0
blinker=1.7.0=pypi_0
bzip2=1.0.8=h1de35cc_0
c-ares=1.23.0=h10d778d_0
ca-certificates=2023.11.17=h8857fd0_0
cachetools=5.3.2=pypi_0
certifi=2023.11.17=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
comm=0.2.0=pypi_0
debugpy=1.8.0=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
distro=1.8.0=pypi_0
executing=2.0.1=pypi_0
fastjsonschema=2.19.0=pypi_0
fqdn=1.5.1=pypi_0
gflags=2.2.2=hb1e8313_1004
gitdb=4.0.11=pypi_0
gitpython=3.1.40=pypi_0
glog=0.6.0=h8ac2a54_0
h11=0.14.0=pypi_0
httpcore=1.0.2=pypi_0
httpx=0.25.2=pypi_0
icu=73.2=hf5e326d_0
idna=3.6=pypi_0
importlib-metadata=6.11.0=pypi_0
ipykernel=6.27.1=pypi_0
ipython=8.18.1=pypi_0
ipywidgets=8.1.1=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.2=pypi_0
json5=0.9.14=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.20.0=pypi_0
jsonschema-specifications=2023.11.2=pypi_0
jupyter-client=8.6.0=pypi_0
jupyter-core=5.5.0=pypi_0
jupyter-events=0.9.0=pypi_0
jupyter-lsp=2.2.1=pypi_0
jupyter-server=2.12.1=pypi_0
jupyter-server-terminals=0.4.4=pypi_0
jupyterlab=4.0.9=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.25.2=pypi_0
jupyterlab-widgets=3.0.9=pypi_0
krb5=1.21.2=hb884880_0
libabseil=20230802.1=cxx17_h048a20a_0
libarrow=14.0.1=h7e6b3eb_9_cpu
libarrow-acero=14.0.1=hc222712_9_cpu
libarrow-dataset=14.0.1=hc222712_9_cpu
libarrow-flight=14.0.1=ha1803ca_9_cpu
libarrow-flight-sql=14.0.1=h8ec153b_9_cpu
libarrow-gandiva=14.0.1=heeebe7c_9_cpu
libarrow-substrait=14.0.1=h8ec153b_9_cpu
libblas=3.9.0=20_osx64_openblas
libbrotlicommon=1.1.0=h0dc2134_1
libbrotlidec=1.1.0=h0dc2134_1
libbrotlienc=1.1.0=h0dc2134_1
libcblas=3.9.0=20_osx64_openblas
libcrc32c=1.1.2=he49afe7_0
libcurl=8.5.0=h726d00d_0
libcxx=16.0.6=hd57cbcb_0
libedit=3.1.20191231=h0678c8f_2
libev=4.33=h10d778d_2
libevent=2.1.12=ha90c15b_1
libffi=3.4.4=hecd8cb5_0
libgfortran=5.0.0=13_2_0_h97931a8_1
libgfortran5=13.2.0=h2873a65_1
libgoogle-cloud=2.12.0=hc0857f6_4
libgrpc=1.59.3=ha7f534c_0
libiconv=1.17=hac89ed1_0
liblapack=3.9.0=20_osx64_openblas
libllvm15=15.0.7=hbedff68_4
libnghttp2=1.58.0=h64cf6d3_1
libopenblas=0.3.25=openmp_hfef2a42_0
libparquet=14.0.1=h27bd29f_9_cpu
libprotobuf=4.24.4=h0ee05dc_0
libre2-11=2023.06.02=h4694dbf_0
libssh2=1.11.0=hd019ec5_0
libthrift=0.19.0=h064b379_1
libutf8proc=2.8.0=hb7f2c08_0
libxml2=2.12.2=hc0ae0f7_0
libzlib=1.2.13=h8a1eda9_5
llvm-openmp=17.0.6=hb6ac08f_0
lz4-c=1.9.4=hf0c8a7f_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib-inline=0.1.6=pypi_0
mdurl=0.1.2=pypi_0
mistune=3.0.2=pypi_0
nbclient=0.9.0=pypi_0
nbconvert=7.12.0=pypi_0
nbformat=5.9.2=pypi_0
ncurses=6.4=hcec6c5f_0
nest-asyncio=1.5.8=pypi_0
notebook=7.0.6=pypi_0
notebook-shim=0.2.3=pypi_0
numpy=1.26.2=py311h93c810c_0
openai=1.3.8=pypi_0
openssl=3.2.0=hd75f5a5_1
orc=1.9.2=h9ab30d4_0
overrides=7.4.0=pypi_0
packaging=23.2=pypi_0
pandas=2.1.4=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pexpect=4.9.0=pypi_0
pillow=10.1.0=pypi_0
pip=23.3.1=py311hecd8cb5_0
platformdirs=4.1.0=pypi_0
prometheus-client=0.19.0=pypi_0
prompt-toolkit=3.0.41=pypi_0
protobuf=4.25.1=pypi_0
psutil=5.9.6=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=14.0.1=py311h98a0319_9_cpu
pycparser=2.21=pypi_0
pydantic=2.5.2=pypi_0
pydantic-core=2.14.5=pypi_0
pydeck=0.8.1b0=pypi_0
pygments=2.17.2=pypi_0
python=3.11.3=hf27a42d_1
python-dateutil=2.8.2=pypi_0
python-decouple=3.8=pypi_0
python-dotenv=1.0.0=pypi_0
python-json-logger=2.0.7=pypi_0
python_abi=3.11=2_cp311
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.2=pypi_0
re2=2023.06.02=hd34609a_0
readline=8.2=hca72f7f_0
referencing=0.32.0=pypi_0
requests=2.31.0=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rich=13.7.0=pypi_0
rpds-py=0.13.2=pypi_0
send2trash=1.8.2=pypi_0
setuptools=68.0.0=py311hecd8cb5_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
snappy=1.1.10=h225ccf5_0
sniffio=1.3.0=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.41.2=h6c40b1e_0
stack-data=0.6.3=pypi_0
streamlit=1.29.0=pypi_0
tenacity=8.2.3=pypi_0
terminado=0.18.0=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h5d9f67b_0
toml=0.10.2=pypi_0
toolz=0.12.0=pypi_0
tornado=6.4=pypi_0
tqdm=4.66.1=pypi_0
traitlets=5.14.0=pypi_0
types-python-dateutil=2.8.19.14=pypi_0
typing-extensions=4.9.0=pypi_0
tzdata=2023.3=pypi_0
tzlocal=5.2=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.1.0=pypi_0
validators=0.22.0=pypi_0
wcwidth=0.2.12=pypi_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.7.0=pypi_0
wheel=0.41.2=py311hecd8cb5_0
widgetsnbextension=4.0.9=pypi_0
xz=5.4.5=h6c40b1e_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h8a1eda9_5
zstd=1.5.5=h829000d_0
